+++ src/activities/settings/SettingsActivity.h
@@ -9,6 +9,7 @@
 
 #include "CrossPointSettings.h"
 #include "activities/Activity.h"
+#include "components/OptionPopup.h"
 #include "util/ButtonNavigator.h"
 
 enum class SettingType { TOGGLE, ENUM, ACTION, VALUE, STRING, SECTION_HEADER, SUBMENU };
@@ -245,6 +246,8 @@ class SettingsActivity final : public Activity {
   bool quickResumeTimeoutAutoEnabled = false;
   SettingAction activeSubmenu = SettingAction::None;
   SettingAction parentSubmenu = SettingAction::None;
+
+  OptionPopup optionPopup;
 
   static constexpr int categoryCount = 4;
   static const StrId categoryNames[categoryCount];